+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/Settings.java
@@ -35,6 +35,10 @@ public class Settings extends BaseSettings {
     public static final StringSetting DOWNLOAD_VIDEO_QUALITY = new StringSetting("download_video_quality", "auto");
     public static final BooleanSetting DOWNLOAD_ORIGINAL_PHOTOS = new BooleanSetting("download_original_photos", FALSE);
     public static final BooleanSetting DOWNLOAD_AUDIO_TRACK = new BooleanSetting("download_audio_track", FALSE);
+    public static final BooleanSetting DOWNLOAD_WITHOUT_SOUND =
+            new BooleanSetting("download_without_sound", FALSE);
+    public static final StringSetting EXTERNAL_DOWNLOADER_PACKAGE =
+            new StringSetting("external_downloader_package", "");
     public static final StringSetting DOWNLOAD_STICKER_FORMAT = new StringSetting("download_sticker_format", "mp4");
     public static final BooleanSetting SAVE_PROFILE_PICTURE = new BooleanSetting("save_profile_picture", FALSE);
     public static final BooleanSetting SAVE_STORY = new BooleanSetting("save_story", FALSE);